+++ karma.conf.js
@@ -1,5 +1,15 @@
-// Karma configuration
-// Generated on Tue Jun 03 2014 12:40:44 GMT+0200 (CEST)
+var _ = require('underscore');
+
+
+var testFiles = [
+    'karma.header.js',
+    'public/scripts/**/tests/**/*.js'
+];
+
+var preprocessors = _.reduce(testFiles, function(memo, file) {
+    memo[file] = ['browserify'];
+    return memo;
+}, {});
 
 module.exports = function(config) {
   config.set({
@@ -10,29 +20,29 @@ module.exports = function(config) {
 
     // frameworks to use
     // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-    frameworks: ['jasmine', 'requirejs'],
+    frameworks: ['browserify', 'jasmine'],
 
 
     // list of files / patterns to load in the browser
-    files: [
-      'public/scripts/require.config.js',
-      {pattern: 'public/scripts/common/lib/filetypes/test-sequences/*', included: false},
-      {pattern: 'public/**/*.js', included: false},
-      {pattern: 'tests/specs/**/*.js', included: false},
-      'test-main.js'
-    ],
+    files: testFiles,
 
 
     // list of files to exclude
     exclude: [
-      'public/scripts/app.js'
     ],
 
 
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
-    preprocessors: {
-    
+    preprocessors: preprocessors,
+
+    browserify: {
+        debug: true,
+        transform: [
+            ['hbsfy', { compiler: 'require("handlebars.mixed");'}],
+            'babelify',
+            'deamdify'
+        ]
     },
