+++ karma.header.js
@@ -1,0 +1,4 @@
+// This file is used to load library mixins, which is usually done in app.js
+
+window._ = require('underscore.mixed');
+var Backbone = require('backbone.mixed');
