+++ package.json
@@ -73,11 +73,15 @@
   "devDependencies": {
     "amdefine": "~0.1.0",
     "benchmark": "~1.0.0",
+    "karma": "^0.12.31",
+    "karma-browserify": "^4.0.0",
+    "karma-phantomjs-launcher": "^0.1.4",
     "pretty-error": "^1.1.1"
   },
   "scripts": {
     "start": "nf start webdev,doc",
     "build": "./node_modules/.bin/gulp build",
+    "test": "./node_modules/.bin/karma start --single-run",
     "watch": "./node_modules/.bin/gulp",
     "postinstall": "npm run build"
   }
