--- test-main.js
@@ -1,19 +1,0 @@
-var tests = [];
-for (var file in window.__karma__.files) {
-  if (window.__karma__.files.hasOwnProperty(file)) {
-    if (/spec\.js$/.test(file)) {
-      tests.push(file);
-    }
-  }
-}
-
-require.config({
-  // Karma serves files under /base, which is the basePath from your config file
-  baseUrl: '/base/public/scripts',
-
-  // dynamically load all test files
-  deps: tests,
-
-  // we have to kickoff jasmine, as it is asynchronous
-  callback: window.__karma__.start
-});
